write stacktrace when panic occured
